requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.1.1
Faker==15.3.3
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.1
pluggy==1.0.0
PyJWT==2.6.0
pymongo==4.3.2
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
